+++ package.json
@@ -37,7 +37,7 @@
     "superjson": "^2.2.1",
     "tailwind-merge": "^2.2.1",
     "zod": "^3.22.4",
-    "zod-prisma-types": "^3.1.6"
+    "zod-prisma-types": "^3.1.1"
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
