+++ schemas/inputTypeSchemas/EventParticipantsScalarFieldEnumSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const EventParticipantsScalarFieldEnumSchema = z.enum(['id','userId','eventId','createdAt']);
+
+export default EventParticipantsScalarFieldEnumSchema;
