+++ schemas/inputTypeSchemas/NullsOrderSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const NullsOrderSchema = z.enum(['first','last']);
+
+export default NullsOrderSchema;
