+++ schemas/inputTypeSchemas/RoleSchema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export const RoleSchema = z.enum(['USER','ADMIN']);
+
+export type RoleType = `${z.infer<typeof RoleSchema>}`
+
+export default RoleSchema;
