+++ schemas/inputTypeSchemas/SessionScalarFieldEnumSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const SessionScalarFieldEnumSchema = z.enum(['id','sessionToken','userId','expires']);
+
+export default SessionScalarFieldEnumSchema;
