+++ schemas/inputTypeSchemas/SortOrderSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const SortOrderSchema = z.enum(['asc','desc']);
+
+export default SortOrderSchema;
