+++ schemas/inputTypeSchemas/TransactionIsolationLevelSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const TransactionIsolationLevelSchema = z.enum(['ReadUncommitted','ReadCommitted','RepeatableRead','Serializable']);
+
+export default TransactionIsolationLevelSchema;
