+++ schemas/inputTypeSchemas/VerificationTokenScalarFieldEnumSchema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const VerificationTokenScalarFieldEnumSchema = z.enum(['identifier','token','expires']);
+
+export default VerificationTokenScalarFieldEnumSchema;
