+++ schemas/modelSchema/UserSchema.ts
@@ -1,0 +1,23 @@
+import { z } from 'zod';
+import { RoleSchema } from '../inputTypeSchemas/RoleSchema'
+
+/////////////////////////////////////////
+// USER SCHEMA
+/////////////////////////////////////////
+
+export const UserSchema = z.object({
+  role: RoleSchema,
+  id: z.string().cuid(),
+  name: z.string().nullable(),
+  username: z.string().nullable(),
+  password: z.string().nullable(),
+  firstName: z.string().nullable(),
+  lastName: z.string().nullable(),
+  email: z.string().nullable(),
+  emailVerified: z.coerce.date().nullable(),
+  image: z.string().nullable(),
+})
+
+export type User = z.infer<typeof UserSchema>
+
+export default UserSchema;
