+++ schemas/modelSchema/VerificationTokenSchema.ts
@@ -1,0 +1,15 @@
+import { z } from 'zod';
+
+/////////////////////////////////////////
+// VERIFICATION TOKEN SCHEMA
+/////////////////////////////////////////
+
+export const VerificationTokenSchema = z.object({
+  identifier: z.string(),
+  token: z.string(),
+  expires: z.coerce.date(),
+})
+
+export type VerificationToken = z.infer<typeof VerificationTokenSchema>
+
+export default VerificationTokenSchema;
