+++ src/components/Event/UpcomingEventTile.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const UpcomingEventTile = () => {
+  return (
+    <div className='flex flex-row basis-1/5'></div>
+  )
+}
+
+export default UpcomingEventTile
